refactor(craft): reorganize build package into domain packages
Routers move into the domain packages that own their logic, matching the
documents/ / query_and_chat/ layout: session/{api,messages,models}.py,
interactive_turns/api.py, user_library/api.py, external_apps/{api,oauth,models}.py,
debug.py. The webapp proxy moves out of the aggregator into webapp_proxy.py,
leaving build/api.py as thin router wiring. api/models.py is split per domain
and five dead classes are removed.
