+++ backend/tests/external_dependency_unit/craft/test_custom_external_app_create.py
@@ -11,16 +11,18 @@ from sqlalchemy import delete
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-import onyx.server.features.build.api.external_apps_api as api
+import onyx.server.features.build.external_apps.api as api
 from onyx.db.enums import ExternalAppType
 from onyx.db.models import ExternalApp
 from onyx.db.models import Skill
 from onyx.db.models import User
 from onyx.error_handling.error_codes import OnyxErrorCode
 from onyx.error_handling.exceptions import OnyxError
-from onyx.server.features.build.api.models import CreateBuiltInExternalAppRequest
-from onyx.server.features.build.api.models import ExternalAppAdminResponse
-from onyx.server.features.build.api.models import UpdateExternalAppRequest
+from onyx.server.features.build.external_apps.models import (
+    CreateBuiltInExternalAppRequest,
+)
+from onyx.server.features.build.external_apps.models import ExternalAppAdminResponse
+from onyx.server.features.build.external_apps.models import UpdateExternalAppRequest
 from onyx.utils.encryption import is_masked_credential
 
 _AUTH_TEMPLATE = {"Authorization": "Bearer {api_key}"}
